[tsock] Implement server link
Updates tsock.sh with a set-server-link command, which is added to tmux's hooks
such that the server link gets updated when the active client changes.

Also sets SSH_AUTH_SOCK to the server link in new in-tmux zsh interactive
shells. Together, this should make agent forwarding "just work", but not fully
tested yet.