[tsock] Run set-server-link as a preexec
The client-active hook doesn't always get the right value of the new
client_tty, so as a workaround we can run set-server-link as a preexec ZSH
hook.  I've updated this function to be more efficient and it doesn't seem to
affect shell responsiveness, but it's still not ideal.

Alternative approaches to consider include:

- Running periodically in the background, either with another tmux hook (status
  line?) or a shell script daemon per tmux server.
- Wrapping the ssh cli and running before opening a connection.