[workflow] Use separate Periodic workflow
On second though, like with coursepointer it's better to use a separate
Periodic workflow that tests against the full rainge of available runner
images.

There's some duplication in here: At the moment ubuntu-latest is the
same as ubuntu-24.04, and macos-latest is macos-14.  But better to do it
this way so that as the -latest images are updated over time and older
images are deprecated, any changes that I need to make are surfaced as
workflow failures.